move auth all to index
ref #61
